Reduced wave empty check to wave length.
The enemy collection lastWaveEnemy event will take care of notifying
the game that the wave is empty and remove it from the game wave set.

Now we only have to check if the length of the wave set is 0 instead of
also checking if each wave in the wave set is empty.