Introduced round tokens.
These tokens are generated by the server and sent to the client. The
round token must be passed from the client back to the server to
execute the command for that token. Once a command has been issued for
the token, it is consumed and can't be used again for this game.